chore: split chat page into parts: components, hooks and utilities
