+++ platform/frontend/src/app/chat/_components/prompt-input/index.tsx
