+++ platform/frontend/src/app/chat/_utils/chat-initial-state.ts
